autostuff: get rid of the bootstrap script
Running autoreconf when AC_CONFIG_MACRO_DIR does not exist ends up in a
failure. Unfortunately, git does not track empty directories. That's why
we have bootstrap, which creates it before running autoreconf.

Add an empty, dummy file in there and track it, so we can get rid of our
bootstrap script.

Update the README file accordingly.

We're left with a stray, generated autom4te.cache file, but that's a
minor annoyance.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>