buildsystem: do not use program-prefix to set the kconfig- prefix
Currently, we use automake's --program-prefix to set the prefix of the
kconfig executables to 'kconfig-'.

However, this is neither very nice nor very reliable. In some cases, the
standard practice distributions use is to force the prefix to be empty,
in which case this would create clashes with other tools (esp. diff and
merge).

Furthermore, we use a dirty trick to achieve this, by initialising
program-prefix before initialising automake. There is no guarantee this
will continue to work with future automake versions.

Stop using this trick, and directly include the 'kconfig-' prefix for
all programs (binaries and scripts alike) that we may generate and
install.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>